Adding mode based execution to the script and restructuring in preparation for packaging
